Separate unit tests into separate executables
I would like to start creating unit tests prior to implementing the
functionality. However, that means the unit test won't build until
the implementation is complete. If the unit tests are separate
executables, I could build all the binaries that are able to be
built (with make -k) and continue to implement the functionality for
the other tests. This will also allow me to get rid of
test/unit_test/main.cpp that likes to clash with the file completion
when tabbing into test/unit_test/math/

Relevant Links:
- Separate unit tests into separate executables
(https://github.com/tybl/math/issues/8)